requirements.txt
fastapi 
uvicorn 
python-multipart 
requests 
python-dotenv 
pillow 
pytesseract 
opencv-python
spacy